+++ ribs/emitters/_gradient_arborescence_emitter.py
@@ -311,8 +311,8 @@ class GradientArborescenceEmitter(EmitterBase):
             return False
         raise ValueError(f"Invalid restart_rule {self._restart_rule}")
 
-    def tell_dqd(self, solution, objective, measures, jacobian, status_batch,
-                 value_batch, **fields):
+    def tell_dqd(self, solution, objective, measures, jacobian, add_info,
+                 **fields):
         """Gives the emitter results from evaluating the gradient of the
         solutions.
 
@@ -328,12 +328,8 @@ class GradientArborescenceEmitter(EmitterBase):
                 from :meth:`ask_dqd`. Each matrix should consist of the
                 objective gradient of the solution followed by the measure
                 gradients.
-            status_batch (array-like): 1d array of
-                :class:`ribs.archive.addstatus` returned by a series of calls
-                to archive's :meth:`add()` method.
-            value_batch (array-like): 1d array of floats returned by a series
-                of calls to archive's :meth:`add()` method. for what these
-                floats represent, refer to :meth:`ribs.archives.add()`.
+            add_info (dict): Data returned from the archive
+                :meth:`~ribs.archives.ArchiveBase.add` method.
             fields (keyword arguments): Additional data for each solution. Each
                 argument should be an array with batch_size as the first
                 dimension.
@@ -346,10 +342,7 @@ class GradientArborescenceEmitter(EmitterBase):
                 "measures": measures,
                 **fields,
             },
-            {
-                "status": status_batch,
-                "value": value_batch,
-            },
+            add_info,
             jacobian,
         )
 
@@ -359,8 +352,7 @@ class GradientArborescenceEmitter(EmitterBase):
             jacobian /= norms
         self._jacobian_batch = jacobian
 
-    def tell(self, solution, objective, measures, status_batch, value_batch,
-             **fields):
+    def tell(self, solution, objective, measures, add_info, **fields):
         """Gives the emitter results from evaluating solutions.
 
         The solutions are ranked based on the `rank()` function defined by
@@ -373,12 +365,8 @@ class GradientArborescenceEmitter(EmitterBase):
                 value of each solution.
             measures (array-like): (batch_size, measure space dimension) array
                 with the measure space coordinates of each solution.
-            status_batch (array-like): 1d array of
-                :class:`ribs.archive.addstatus` returned by a series of calls
-                to archive's :meth:`add()` method.
-            value_batch (array-like): 1d array of floats returned by a series
-                of calls to archive's :meth:`add()` method. for what these
-                floats represent, refer to :meth:`ribs.archives.add()`.
+            add_info (dict): Data returned from the archive
+                :meth:`~ribs.archives.ArchiveBase.add` method.
             fields (keyword arguments): Additional data for each solution. Each
                 argument should be an array with batch_size as the first
                 dimension.
@@ -394,10 +382,7 @@ class GradientArborescenceEmitter(EmitterBase):
                 "measures": measures,
                 **fields,
             },
-            {
-                "status": status_batch,
-                "value": value_batch,
-            },
+            add_info,
         )
 
         if self._jacobian_batch is None:
@@ -408,7 +393,7 @@ class GradientArborescenceEmitter(EmitterBase):
         self._itrs += 1
 
         # Count number of new solutions.
-        new_sols = status_batch.astype(bool).sum()
+        new_sols = add_info["status"].astype(bool).sum()
 
         # Sort the solutions using ranker.
         indices, ranking_values = self._ranker.rank(self, self.archive,